Make version_string() public
This allows downstream crates to access it, and write the string
to their own CLIs and metadata files
